Fixed config file to have data types.
Now datatypes will be saved and enforced,
this one fixes the case that password is only numbers
so it will be read as int, and then you are stuck out since
no metter how much times you gonna enter the pass you will fail.
fixing this will enforce datatypes to be same as set in configs.

after applying this, deleting the old config file is a must!