Handle permission error in shutil.move for *nix systems during the rename plugin.
shutil.move automatically calls shutil.copystat after moving a file.  On a *nix system copystat tries to update file permissions of the new file.  If the destination has an NTFS file system the chmod will fail with a simple permission error.  This commit checks for this specific condition and if the destination file exists supresses the error and deletes the original file.