Update main.py
This the change in response to issue Issue #1439. The imdb identifier search doesn't work anymore on nzbsrus, so I changed the search string to only the title of the movie.

I replaced question marks in the title because the api won't take them.

I know it's probably not the best solution, suggestions for a better solution are welcome.