Additional option to work around permissions issues
When couchpotato creates the folder for the torrent file, it does so under whatever user it is running as. This can create problems with transmission, which by default runs as its own user. As a workaround for this, I propose setting an option that makes couchpotato behave like other content searchers (e.g. sickbeard) in that it simply lets the torrent app handle folder creation.