Adjust force renamer scan
The force scan function causes trouble for several users. For example it
also picks up seeding torrents and may also pick up incomplete torrents
if there has been no downloading activity for > 1 minute.

This commit changes the following items:
- Disable it by default
- Only scan at start-up if the force scan is enabled (>0)
- Add a warning in settings that it can mess up torrents

We could remove the force scan function all together by simply adding a
manual button to scan the From folder. Note that if we do this, warnings
should be added that it also picks up seeding and potentially incomplete
torrents.

As a last note: there is no way to see from the file if it has completed
downloading or seeding! The only way is to ask the downloader. This
could mean that seeding or incomplete files should not be in the from
folder in the first place?