Allow BRRip for 1080p and 720p
While trying to download from my preferred group which generally uses
720p.BRRip in the name, their releases wouldn't be chosen even with
the group name in my preferred words. This was because 720p quality
doesn't match anything with any other quality identifier in it.
Since bluray is essentially the highest quality source media, there
is no reason not to allow it in combination with 1080p and 720p.