Fixed search criteria to work with multi-word titles.
Search string as composed before was in form ["title" year]. But this did not work for longer titles - for example for After Earth search string was ["After Earth" 2013]. TorrentShack then required the releases to have space in name and most of relevant results were filtered out (they have dot in name). Changed search to replace space in the title with underscore which works as a wildchar. So the search string is now ["After_Earth" 2013].