Update README.md
I am on Manjaro linux. I dowloaded and installed CP according to the instructions on this `python CouchPotatoServer/CouchPotato.py` but it didn't run.

Since this requires python2 the script wont run unless it is edited at the top from 

> #!/usr/bin/env python

to `#!/usr/bin/python2`
and the command to enter in the terminal is `python2 CouchPotatoServer/CouchPotato.py` 

I realized this after wondering why it wasn't running.