Extend ignore patterns
Some want to synchronize files between a download server and storage server
hosting the couch potato instance. The renaming plugin does copy files when
file beeing available regardless the synchronization status, resulting in
corrupted files in destination folder. Files are not copied if a file with
extension .ignored is present. This extend at the moment the ignored extensions
to support common lftp temporary file. This way files get only copied once
suchronization is ended.
In the future this could be extended to have this extension list a configurable
option in the settings.