Switch from Ruby Sass to node-sass
Ruby Sass is officially deprecated (https://sass-lang.com/ruby-sass).
This commit switches the Sass implementation from Ruby Sass to node-sass
(which is a wrapper around libsass).

Ruby Sass used config.rb to configure some paths; as this is not
supported by node-sass, I needed to change the import statements in the
scss files back to explicit paths.

Also updated the README file to reflect this (and added an additional
section about testing while I was at it).