Use 7 letter IMDB ID's unless title has an 8 char long IMDB ID.
IMDB still returns 7-char ID's on older titles when searching, simply
blanketin to 8 chars all the time (even if IMDB support 8-char padded
ID's) will ultimately fail lookups on other services.

This allows us to default to a 7-char ID, padded with zero's to ensure
that it still works as expected, but ultimately allow newer 8-char ID's
as-and-when required for neweer titles.